# scoreboard-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/shotclock/set": {
            "post": {
                "description": "Sets the shot clock to a specific value in ss.x format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shotclock"
                ],
                "summary": "Set the shot clock",
                "parameters": [
                    {
                        "description": "Shot clock in ss.x format",
                        "name": "shotclock",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SetShotClockRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/timer/set": {
            "post": {
                "description": "Sets the main timer to a specific value in mm:ss format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timer"
                ],
                "summary": "Set the timer",
                "parameters": [
                    {
                        "description": "Timer in mm:ss format",
                        "name": "timer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SetTimerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.SetShotClockRequest": {
            "description": "Timer in ss.x format",
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                }
            }
        },
        "handlers.SetTimerRequest": {
            "description": "Timer in mm:ss format",
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                }
            }
        }
    }
}